Add files via upload
This genshin and project.html are combined project. This project is about a game called Genshin impact. I have created a redeemption site prototype.
YOU have tried to sign in amazon for a purchase, it's just like that .

I you guys have any suggestions on making the site wonderful and functionable. PLEASE contribute to my project.